+++ src/Data/Vector/Generic/Lens.hs
@@ -4,6 +4,10 @@
 #ifdef TRUSTWORTHY
 {-# LANGUAGE Trustworthy #-}
 #endif
+
+#ifndef MIN_VERSION_vector
+#define MIN_VERSION_vector(x,y,z) 1
+#endif
 -------------------------------------------------------------------------------
 -- |
 -- Module      :  Data.Vector.Generic.Lens
@@ -45,9 +49,14 @@ import Control.Lens.Traversal
 import Control.Lens.Internal.List (ordinalNub)
 import Data.Monoid
 import Data.Vector.Generic as V hiding (zip, filter, indexed)
-import Data.Vector.Fusion.Stream (Stream)
 import Data.Vector.Generic.New (New)
 import Prelude hiding ((++), length, null, head, tail, init, last, map, reverse)
+
+#if MIN_VERSION_vector(0,11,0)
+import Data.Vector.Fusion.Bundle (Bundle)
+#else
+import Data.Vector.Fusion.Stream (Stream)
+#endif
 
 -- $setup
 -- >>> import Data.Vector as Vector
@@ -92,14 +101,25 @@ vector :: (Vector v a, Vector v b) => Iso [a] [b] (v a) (v b)
 vector = iso fromList V.toList
 {-# INLINE vector #-}
 
+#if MIN_VERSION_vector(0,11,0)
+-- | Convert a 'Vector' to a finite 'Bundle' (or back.)
+asStream :: (Vector v a, Vector v b) => Iso (v a) (v b) (Bundle v a) (Bundle v b)
+#else
 -- | Convert a 'Vector' to a finite 'Stream' (or back.)
 asStream :: (Vector v a, Vector v b) => Iso (v a) (v b) (Stream a) (Stream b)
+#endif
 asStream = iso stream unstream
 {-# INLINE asStream #-}
 
+#if MIN_VERSION_vector(0,11,0)
+-- | Convert a 'Vector' to a finite 'Bundle' from right to left (or
+-- back.)
+asStreamR :: (Vector v a, Vector v b) => Iso (v a) (v b) (Bundle v a) (Bundle v b)
+#else
 -- | Convert a 'Vector' to a finite 'Stream' from right to left (or
 -- back.)
 asStreamR :: (Vector v a, Vector v b) => Iso (v a) (v b) (Stream a) (Stream b)
+#endif
 asStreamR = iso streamR unstreamR
 {-# INLINE asStreamR #-}